Rename `path` to `temp_fs`
